+++ packages/components/src/components/popover/popover.tsx
@@ -1,14 +1,19 @@
-import { Component, Element, h, Host, type JSX, Prop, State } from '@stencil/core';
+import { Component, Element, forceUpdate, h, Host, type JSX, Prop, State } from '@stencil/core';
+import type { PopoverAriaAttribute, PopoverDirection } from './popover-utils';
 import {
   addDocumentEventListener,
+  addNativeScrollAndResizeListeners,
   POPOVER_ARIA_ATTRIBUTES,
   POPOVER_DIRECTIONS,
   removeDocumentEventListener,
+  updateNativePopoverStyles,
   updatePopoverStyles,
 } from './popover-utils';
 import {
   AllowedTypes,
   attachComponentCss,
+  findClosestComponent,
+  getHasNativePopoverSupport,
   getPrefixedTagNames,
   hasPropValueChanged,
   parseAndGetAriaAttributes,
@@ -16,7 +21,6 @@ import {
   validateProps,
 } from '../../utils';
 import { getComponentCss } from './popover-styles';
-import type { PopoverAriaAttribute, PopoverDirection } from './popover-utils';
 import type { PropTypes, SelectedAriaAttributes, Theme } from '../../types';
 
 const propTypes: PropTypes<typeof Popover> = {
@@ -52,8 +56,12 @@ export class Popover {
   private popover: HTMLDivElement;
   private button: HTMLButtonElement;
 
+  private isNative: boolean = false;
+  private parentTableElement: HTMLElement;
+
   public connectedCallback(): void {
     addDocumentEventListener(this);
+    this.detectNativeCase();
   }
 
   public componentShouldUpdate(newVal: unknown, oldVal: unknown): boolean {
@@ -61,9 +69,17 @@ export class Popover {
   }
 
   public componentDidRender(): void {
-    if (this.open) {
-      // calculate / update position only possible after render
-      updatePopoverStyles(this.host, this.spacer, this.popover, this.direction, this.theme);
+    if (this.isNative && this.spacer?.matches(':popover-open')) {
+      addNativeScrollAndResizeListeners(this.host, this.parentTableElement, this.spacer);
+      // Set new popover position depending on button position
+      updateNativePopoverStyles(this.spacer, this.button);
+      // Update popover styles with new position
+      updatePopoverStyles(this.host, this.spacer, this.popover, this.direction, this.isNative, this.theme);
+    } else {
+      if (this.open) {
+        // calculate / update position only possible after render
+        updatePopoverStyles(this.host, this.spacer, this.popover, this.direction, false, this.theme);
+      }
     }
   }
 
@@ -73,7 +89,7 @@ export class Popover {
 
   public render(): JSX.Element {
     validateProps(this, propTypes);
-    attachComponentCss(this.host, getComponentCss, this.direction, this.theme);
+    attachComponentCss(this.host, getComponentCss, this.direction, this.isNative, this.theme);
 
     const PrefixedTagNames = getPrefixedTagNames(this.host);
 
@@ -81,18 +97,22 @@ export class Popover {
       <Host onKeydown={this.onKeydown}>
         <button
           type="button"
-          onClick={() => (this.open = !this.open)}
+          {...(this.isNative ? { popoverTarget: 'spacer' } : { onClick: () => (this.open = !this.open) })}
           {...parseAndGetAriaAttributes({
             ...parseAndGetAriaAttributes(this.aria),
-            'aria-expanded': this.open,
+            ...(!this.isNative && { 'aria-expanded': this.open }),
           })}
           ref={(el) => (this.button = el)}
         >
           <PrefixedTagNames.pIcon class="icon" name="information" theme={this.theme} />
           <span class="label">More information</span>
         </button>
-        {this.open && (
-          <div class="spacer" ref={(el) => (this.spacer = el)}>
+        {(this.open || this.isNative) && (
+          <div
+            class="spacer"
+            ref={(el) => (this.spacer = el)}
+            {...(this.isNative && { popover: 'auto', id: 'spacer', onToggle: this.onToggle })}
+          >
             <div class="popover" ref={(el) => (this.popover = el)}>
               {this.description ? <p>{this.description}</p> : <slot />}
             </div>
@@ -105,6 +125,21 @@ export class Popover {
   private onKeydown = (e: KeyboardEvent): void => {
     if (e.key === 'Escape') {
       this.button.focus();
+    }
+  };
+
+  private detectNativeCase = (): void => {
+    if (getHasNativePopoverSupport()) {
+      this.parentTableElement = findClosestComponent(this.host, 'pTable');
+      if (!!this.parentTableElement) {
+        this.isNative = true;
+      }
+    }
+  };
+
+  private onToggle = (e): void => {
+    if (e.newState === 'open') {
+      forceUpdate(this.host); // Necessary to update popover styles since opening of native popover doesn't trigger rerender
     }
   };
 }